Renamed SimilaritySearch → VectorSimilarity and clarified cosine metrics
